Now rebuilds site on changes (first cut)

+++ lesstatic.py
@@ -1,6 +1,6 @@
-import config 
+import config
 import process_source as ps
-import os
+import os,sys
 import shutil
 import re,time
 
@@ -8,7 +8,8 @@ def build_site():
     if os.path.exists(config.site):
         shutil.rmtree(config.site)
     os.mkdir(config.site)
-    shutil.copytree(config.static,config.site+"/"+config.static_dest)
+    if os.path.exist(config.static):
+        shutil.copytree(config.static,config.site+"/"+config.static_dest)
 
     valid_extensions = config.extensions.keys()
     valid_extensions.append('html')
@@ -27,14 +28,22 @@ def build_site():
                 outfile.close()
 
 
-
 def serve():
-    build_site()
+    olds=" ".join([" ".join(["%f"%os.stat(dir+"/"+f).st_mtime for f in files if f[0]!='.']) for (dir,subs,files) in os.walk(config.content)])
+    while  True:
+        time.sleep(1)
+        news=" ".join([" ".join(["%f"%os.stat(dir+"/"+f).st_mtime for f in files if f[0] !='.']) for (dir,subs,files) in os.walk(config.content)])
+        if news!=olds:
+            print "REBUILDING"
+            build_site()
+            olds = news
+
+
 
 if __name__=="__main__":
-    config.base_dir=os.getcwd()
-    print "HELLO"
-    serve()
-
-
-
+    if len(sys.argv)>1:
+        os.chdir(sys.argv[-1])
+        config.base_dir=os.getcwd()
+        build_site()
+    if sys.argv[1]=='serve':
+        serve()
